Set 'biber' as bibliography backend in the example
Currently, the default value in the class file is still 'bibtex8' in order not to break code of students when they just update the class file. However, we already specify 'biber' in the example file so that students will hopefully use 'biber'. In a few months, we should adjust the class file accordingly to finish the transition to biber.