Changes to the construction of the 'Kopfzeile'
By constructing the header manually (see commit #23f8598), we started to rely on the TU fonts to be installed (in order to print the faculty and institute name). As this can not be guaranteed (especially for students), this commit changes the manual construction so that a (separate) pdf 'IfA_Header' containing the faculty and institute name is included instead.